+++ src/app/(auth)/auth/sign-in/page.tsx
@@ -1,13 +1,80 @@
+"use client";
+
 import React from "react";
-import Login from "../../components/login";
+import { FormEvent } from "react";
+import { Input } from "@/components/ui/input";
+import { Button } from "@/components/ui/button";
+import PasswordInput from "@/components/ui/password-input";
+import { toast } from "@/components/ui/use-toast";
+import { ROUTES } from "@/utils/constants";
+import { createClient } from "@/utils/supabase/client";
+import cookie from "js-cookie";
+import Link from "next/link";
 
 export default function Page() {
+  const handleFormSubmit = async (e: FormEvent) => {
+    e.preventDefault();
+
+    const formData = new FormData(e.target as HTMLFormElement);
+    const email = formData.get("email")!.toString();
+    const password = formData.get("password")!.toString();
+
+    const supabase = createClient();
+
+    const { error: signInError } = await supabase.auth.signInWithPassword({
+      email,
+      password,
+    });
+
+    if (signInError) {
+      return toast({
+        description: signInError.message,
+      });
+    }
+
+    let { data: users, error } = await supabase
+      .from("users")
+      .select()
+      .eq("email", email);
+
+    if (error || !users) {
+      return toast({
+        description: error?.message,
+      });
+    }
+
+    cookie.set("db_user", JSON.stringify(users[0]));
+
+    window.location.href = ROUTES.overview;
+  };
   return (
-    <div className="p-10">
-      <div className="max-w-[500px] mx-auto">
-        <h1 className="text-2xl">Login to Onena</h1>
-        <Login />
-      </div>
-    </div>
+    <>
+      <h1 className="text-3xl">Login to Onena</h1>
+      <form className="w-full space-y-4" onSubmit={handleFormSubmit}>
+        <label htmlFor="">Email</label>
+        <Input type="email" id="email" name="email" required />
+        <div className="w-full space-y-2">
+          <label htmlFor="">Password</label>
+          <PasswordInput
+            key="password"
+            id="password"
+            name="password"
+            required
+          />
+        </div>
+        <div className="flex items-center justify-between">
+          <Button>Continue</Button>
+          <Link className="underline" href={ROUTES.auth.forgotPassword}>
+            Forgot password?
+          </Link>
+        </div>
+      </form>
+      <p>
+        Don&apos;t have an account?&nbsp;
+        <Link className="underline" href={ROUTES.auth.signUp}>
+          Sign up
+        </Link>
+      </p>
+    </>
   );
 }
